Optimized push relabel + benchmark

+++ Max flow/algorithms/counter.cpp
@@ -1,5 +1,5 @@
 namespace C {
-    int ff_edges_visited = 0;
-    int ek_edges_visited = 0;
-    int dinic_edges_visited = 0;
+    unsigned long long ff_edges_visited = 0;
+    unsigned long long ek_edges_visited = 0;
+    unsigned long long dinic_edges_visited = 0;
 }
