+++ Max flow/algorithms/counter.h
@@ -2,9 +2,9 @@
 #define COUNTER_H
 
 namespace C {
-    extern int ff_edges_visited;
-    extern int ek_edges_visited;
-    extern int dinic_edges_visited;
+    extern unsigned long long ff_edges_visited;
+    extern unsigned long long ek_edges_visited;
+    extern unsigned long long dinic_edges_visited;
 }
 
 #endif
